Promotes playrho::d2::LimitState to playrho::LimitState
