+++ Library/CMakeLists.txt
@@ -77,7 +77,6 @@ set(PLAYRHO_D2_HDRS
 	"${PLAYRHO_INCLUDE_PREFIX}d2/IndexPair.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/Joint.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/JointConf.hpp"
-	"${PLAYRHO_INCLUDE_PREFIX}d2/LimitState.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/Manifold.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/MassData.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/MotorJointConf.hpp"
@@ -145,7 +144,6 @@ set(PLAYRHO_D2_SRCS
 	"${PLAYRHO_SOURCE_PREFIX}d2/GearJointConf.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/Joint.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/JointConf.cpp"
-	"${PLAYRHO_SOURCE_PREFIX}d2/LimitState.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/Manifold.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/MassData.cpp"
 	"${PLAYRHO_SOURCE_PREFIX}d2/MotorJointConf.cpp"
