+++ Library/include/playrho/LimitState.hpp
@@ -19,25 +19,33 @@
  * 3. This notice may not be removed or altered from any source distribution.
  */
 
-#include <playrho/d2/LimitState.hpp>
-
-#include <cassert>
+#ifndef PLAYRHO_D2_LIMITSTATE_HPP
+#define PLAYRHO_D2_LIMITSTATE_HPP
 
 namespace playrho {
-namespace d2 {
 
-const char* ToString(LimitState val) noexcept
+/// @brief Limit state.
+/// @note Only used by joints that implement some notion of a limited range.
+enum class LimitState
 {
-    switch (val)
-    {
-        case LimitState::e_atLowerLimit: return "at lower";
-        case LimitState::e_atUpperLimit: return "at upper";
-        case LimitState::e_equalLimits: return "equal";
-        case LimitState::e_inactiveLimit: break;
-    }
-    assert(val == LimitState::e_inactiveLimit);
-    return "inactive";
-}
+    /// @brief Inactive limit.
+    e_inactiveLimit,
 
-} // namespace d2
+    /// @brief At-lower limit.
+    e_atLowerLimit,
+
+    /// @brief At-upper limit.
+    e_atUpperLimit,
+
+    /// @brief Equal limit.
+    /// @details Equal limit is used to indicate that a joint's upper and lower limits
+    ///   are approximately the same.
+    e_equalLimits
+};
+
+/// @brief Provides a human readable C-style string uniquely identifying the given limit state.
+const char* ToString(LimitState val) noexcept;
+
 } // namespace playrho
+
+#endif // PLAYRHO_D2_LIMITSTATE_HPP
