+++ Library/include/playrho/d2/RevoluteJointConf.hpp
@@ -22,11 +22,11 @@
 #ifndef PLAYRHO_D2_REVOLUTEJOINTCONF_HPP
 #define PLAYRHO_D2_REVOLUTEJOINTCONF_HPP
 
+#include <playrho/LimitState.hpp>
 #include <playrho/Math.hpp>
 #include <playrho/Span.hpp>
 
 #include <playrho/d2/JointConf.hpp>
-#include <playrho/d2/LimitState.hpp>
 
 namespace playrho {
