+++ Library/include/playrho/d2/RopeJointConf.hpp
@@ -22,10 +22,11 @@
 #ifndef PLAYRHO_D2_ROPEJOINTCONF_HPP
 #define PLAYRHO_D2_ROPEJOINTCONF_HPP
 
-#include <playrho/d2/JointConf.hpp>
+#include <playrho/LimitState.hpp>
 #include <playrho/Math.hpp>
 #include <playrho/Span.hpp>
-#include <playrho/d2/LimitState.hpp>
+
+#include <playrho/d2/JointConf.hpp>
 
 namespace playrho {
